add test for the get user endpoint and create user

+++ storage/TempStorage.go
@@ -3,7 +3,7 @@ package storage
 type TempStore interface {
 	GetVal(key string) string
 	SetKeyToVal(key string, value string) bool
-	SetKeyToValWIthExpiry(key string, value string, expiry int) bool
-	ChangeKeyEpiry(key string, newExpiry int) bool
+	SetKeyToValWIthExpiry(key string, value string, expiry float64) bool
+	ChangeKeyEpiry(key string, newExpiry float64) bool
 	DelKey(key string) bool
 }
